[Fix #25] Fixing validation date offset
Currently, validation dates on transactions are stored in the French
timezone, but in Odoo the timezone is expected to be in UTC.
A minor fix is needed in the _payfip_evaluate_data method to
convert the timezone into a UTC one.

Migrate all previous dates validations to fix the timezone issue.
All transactions are processed in chunk of one hundred.

Signed-off-by: Logan Gonet <logan.gonet@horanet.com>