[Fix #15] Prevent ValueError when localizing a non-naive datetime object with pytz (#35)
The previous implementation attempted to localize a datetime object
that was already initialized with timezone information
(`tzinfo=payfip_tz`). This is an incorrect usage of `pytz` and
results in a `ValueError`.

This commit modifies the code to first create a naive datetime
object (without explicit timezone information) and then use
`payfip_tz.localize()` to correctly interpret it within the
payfip timezone.
This resolves the `ValueError` and ensures the subsequent
conversion to UTC is based on the intended timezone.

Signed-off-by: Logan Gônet <logan.gonet@horanet.com>